requirements.txt
discord.py>=2.0.0
python-dotenv>=0.20.0
yt-dlp>=2023.3.4
validators>=0.20.0
PyNaCl>=1.4.0 